bioconda-backup-old/requirements.txt
docker
requests
PyYAML
jinja2
anchorecli
